Update knives.dm
For some reason the psydonite knives are much better than other knives. Their greatest asset is that they are silver. They should not be stronger then steel or iron, since the trope of witchers and others who use silver weapons to hunt monsters is that it is weaker than iron, but the "weakness" of monsters to silver makes them worthwhile. I think making the silver throwdaggers better than even the iron and steel ones to kill normal mortals goes against that logic and just plainly makes them unjustifiably stronger.