Commented-Out Duplicate Clothing Files.
Initial Commit.
Duplicate files for clothing items.
All fields currently commented out: this is temporary.
Entries will be culled down to just names and descs, plus any local balance changes.
As objects are edited, the excess details will be culled.
This is to remove the enormous amount of 'Azuria Localization' in the names and descriptions that currently exist.
All files have their duplicate added so future additions can be added in here, regardless if localization changes were needed to be made.